+++ src/disj.h
@@ -1,7 +1,7 @@
 #ifndef DISJ_H_
 #define DISJ_H_
 
-#include <stdint.h>
+
 
 
 
@@ -9,14 +9,14 @@
 /*Disjoint data set - warunek nie moze byc cykli w grafie*/
 
 struct disj_set{
-	uint16_t parent;
-	uint16_t rnk;
+	int parent;
+	int rnk;
 };
 
  
 
 int compareEdges(const void* edge_1, const void* edge_2);
-uint16_t disj_find(struct disj_set* ssets, uint16_t vertice);
+int disj_find(struct disj_set* ssets, int vertice);
